brickton/requirements.txt
PyQt=>5.0
